refactor: 검수 Entity에서 UserId 대신 User 외래키로 변경 및 UI에서 UserId 대신 UserName이 표기 되도록 변경

+++ src/main/java/com/ohgiraffers/warehousemanagement/wms/inspection/repository/InspectionRepository.java
@@ -5,13 +5,21 @@ import com.ohgiraffers.warehousemanagement.wms.inspection.model.common.Inspectio
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
 public interface InspectionRepository extends JpaRepository<Inspection, Long> {
-    Optional<Inspection> findByTransactionTypeAndTransactionId(InspectionTransactionType transactionType, Integer transactionId);
+    Optional<Inspection> findByTransactionTypeAndTransactionId(InspectionTransactionType transactionType, Long transactionId);
 
     Page<Inspection> findAllByOrderByInspectionIdDesc(Pageable pageable);
 
     Page<Inspection> findAllByTransactionTypeOrderByInspectionIdDesc(InspectionTransactionType transactionType, Pageable pageable);
+
+//    @Query("SELECT i FROM Inspection i WHERE " +
+//            "(:searchType = 'inspectionId' AND i.inspectionId LIKE %:search%) OR " +
+//            "(:searchType = 'userId' AND i.userId LIKE %:search%) OR " +
+//            "(:searchType = 'transactionId' AND i.transactionId LIKE %:search%) OR " +
+//            "(:searchType = 'inspectionDate' AND i.inspectionDate LIKE :search)")
+//    Page<Inspection> findAllBySearchOrderByInspectionIdDesc(String searchType, Long search, Pageable pageable);
 }
